+++ Engine/Source/Runtime/Primitive/Intersector.h
@@ -1,24 +1,42 @@
 #pragma once
+#include "Line3.h"
+#include "Plane3.h"
+#include "Triangle3.h"
+#include "Triangle3.h"
+#include "Sphere3.h"
+#include "OBB3.h"
+#include "AABB3.h"
+#include "Polygon3.h"
 
-namespace Matirx
+namespace Matrix
 {
 	namespace Primitive
 	{
+		//设计基类预留设计
+		template<typename primitive1, typename pritive2>
 		class Intersector
 		{
 		public:
-			//判断两个物体的位置关系的，通过其英文名称都可以判断出它们的具体含义
-#define VSFRONT 0
-#define VSBACK 1
-#define VSON 2
-#define VSCLIPPED 3
-#define VSCULLED 4
-#define VSVISIBLE 5
-#define VSINTERSECT 3
-#define VSOUT 4
-#define VSIN 5
-#define VSNOINTERSECT 6
-//
+
 		};
+
+		template<typename T2>
+		class Intersector<Line3, T2>
+		{
+		public:
+			int RelationWith(const Triangle3& Triangle, bool bCull, VSREAL& fLineParameter, VSREAL fTriangleParameter[3]) const
+			{
+			}
+			/*----------------------------------------------------------------*/
+		};
+
+
+
+
+
+
+
+
+
 	}
 }
