+++ Engine/Source/Runtime/Primitive/OBB3.cpp
@@ -295,3 +295,553 @@ Primitive::AABB3 Matrix::Primitive::OBB3::GetAABB() const
 {
 	return Primitive::AABB3();
 }
+
+
+
+VSREAL OBB3::SquaredDistance(const Matrix::Math::Vector3& Point, VSREAL fOBBParameter[3]) const
+{
+	Matrix::Math::Vector3 Diff = Point - m_Center;
+
+	VSREAL fSqrDistance = 0.0f, fDelta;
+
+	fOBBParameter[0] = Diff.Dot(m_A[0]);
+	if (fOBBParameter[0] < -m_fA[0])
+	{
+		fDelta = fOBBParameter[0] + m_fA[0];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[0] = -m_fA[0];
+	}
+	else if (fOBBParameter[0] > m_fA[0])
+	{
+		fDelta = fOBBParameter[0] - m_fA[0];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[0] = m_fA[0];
+	}
+
+	fOBBParameter[1] = Diff.Dot(m_A[1]);
+	if (fOBBParameter[1] < -m_fA[1])
+	{
+		fDelta = fOBBParameter[1] + m_fA[1];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[1] = -m_fA[1];
+	}
+	else if (fOBBParameter[1] > m_fA[1])
+	{
+		fDelta = fOBBParameter[1] - m_fA[1];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[1] = m_fA[1];
+	}
+
+	fOBBParameter[2] = Diff.Dot(m_A[2]);
+	if (fOBBParameter[2] < -m_fA[2])
+	{
+		fDelta = fOBBParameter[2] + m_fA[2];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[2] = -m_fA[2];
+	}
+	else if (fOBBParameter[2] > m_fA[2])
+	{
+		fDelta = fOBBParameter[2] - m_fA[2];
+		fSqrDistance += fDelta * fDelta;
+		fOBBParameter[2] = m_fA[2];
+	}
+	return fSqrDistance;
+}
+
+VSREAL OBB3::SquaredDistance(const Line3& Line, VSREAL fOBBParameter[3], VSREAL& fLineParameter) const
+{
+	Rectangle3 Rectangle[6];
+	GetRectangle(Rectangle);
+	VSREAL fSqrDist = VSMAX_REAL;
+	VSREAL fSqrDistTmp;
+	VSREAL t, fAABB[3];
+	for (int i = 0; i < 6; i++)
+	{
+		fSqrDistTmp = Line.SquaredDistance(*this, t, fAABB);
+		if (fSqrDist < fSqrDistTmp)
+		{
+			fSqrDist = fSqrDistTmp;
+			fLineParameter = t;
+			for (int j = 0; j < 3; j++)
+			{
+				fOBBParameter[j] = fAABB[j];
+			}
+		}
+	}
+	return fSqrDist;
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const Triangle3& Triangle, VSREAL OBBParameter[3],
+	VSREAL TriangleParameter[3]) const
+{
+
+	Rectangle3 Rectangle[6];
+	GetRectangle(Rectangle);
+	VSREAL fSqrDist = VSMAX_REAL;
+	VSREAL fSqrDistTmp;
+	VSREAL RectP[2], TriP[3];
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[0], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = m_fA[0];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = RectP[1];
+	}
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[1], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = -m_fA[0];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = -RectP[1];
+	}
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[2], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = RectP[1];
+		OBBParameter[1] = m_fA[1];
+		OBBParameter[2] = RectP[0];
+	}
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[3], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = -RectP[1];
+		OBBParameter[1] = -m_fA[1];
+		OBBParameter[2] = RectP[0];
+	}
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[4], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = -RectP[1];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = m_fA[2];
+	}
+
+	fSqrDistTmp = Triangle.SquaredDistance(Rectangle[5], TriP, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		TriangleParameter[0] = TriP[0];
+		TriangleParameter[1] = TriP[1];
+		TriangleParameter[2] = TriP[2];
+		OBBParameter[0] = RectP[1];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = -m_fA[2];
+	}
+	return fSqrDist;
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const Rectangle3& Rectangle, VSREAL OBBParameter[3],
+	VSREAL RectangleParameter[2]) const
+{
+
+	Rectangle3 RectangleTemp[6];
+	GetRectangle(RectangleTemp);
+	VSREAL fSqrDist = VSMAX_REAL;
+	VSREAL fSqrDistTmp;
+	VSREAL RectP[2], RectP1[2];
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[0], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = m_fA[0];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = RectP[1];
+	}
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[1], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = -m_fA[0];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = -RectP[1];
+	}
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[2], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = RectP[1];
+		OBBParameter[1] = m_fA[1];
+		OBBParameter[2] = RectP[0];
+	}
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[3], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = -RectP[1];
+		OBBParameter[1] = -m_fA[1];
+		OBBParameter[2] = RectP[0];
+	}
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[4], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = -RectP[1];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = m_fA[2];
+	}
+
+	fSqrDistTmp = Rectangle.SquaredDistance(RectangleTemp[5], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		RectangleParameter[0] = RectP1[0];
+		RectangleParameter[1] = RectP1[1];
+
+		OBBParameter[0] = RectP[1];
+		OBBParameter[1] = RectP[0];
+		OBBParameter[2] = -m_fA[2];
+	}
+	return fSqrDist;
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const OBB3& OBB, VSREAL OBB1Parameter[3],
+	VSREAL OBB2Parameter[3]) const
+{
+
+	Rectangle3 RectangleTemp[6];
+	GetRectangle(RectangleTemp);
+	VSREAL fSqrDist = VSMAX_REAL;
+	VSREAL fSqrDistTmp;
+	VSREAL RectP[2], RectP1[3];
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[0], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = m_fA[0];
+		OBB1Parameter[1] = RectP[0];
+		OBB1Parameter[2] = RectP[1];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[1], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = -m_fA[0];
+		OBB1Parameter[1] = RectP[0];
+		OBB1Parameter[2] = -RectP[1];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[2], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = RectP[1];
+		OBB1Parameter[1] = m_fA[1];
+		OBB1Parameter[2] = RectP[0];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[3], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = -RectP[1];
+		OBB1Parameter[1] = -m_fA[1];
+		OBB1Parameter[2] = RectP[0];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[4], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = -RectP[1];
+		OBB1Parameter[1] = RectP[0];
+		OBB1Parameter[2] = m_fA[2];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+
+	fSqrDistTmp = OBB.SquaredDistance(RectangleTemp[5], RectP1, RectP);
+	if (fSqrDistTmp < fSqrDist)
+	{
+		fSqrDist = fSqrDistTmp;
+		OBB1Parameter[0] = RectP[1];
+		OBB1Parameter[1] = RectP[0];
+		OBB1Parameter[2] = -m_fA[2];
+
+		OBB2Parameter[0] = RectP1[0];
+		OBB2Parameter[1] = RectP1[0];
+		OBB2Parameter[2] = RectP1[1];
+	}
+	return fSqrDist;
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const Ray3& Ray, VSREAL fOBBParameter[3], VSREAL& fRayParameter) const
+{
+	return Ray.SquaredDistance(*this, fRayParameter, fOBBParameter);
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const Segment3& Segment, VSREAL fOBBParameter[3], VSREAL& fSegmentParameter) const
+{
+	return Segment.SquaredDistance(*this, fSegmentParameter, fOBBParameter);
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::Distance(const Sphere3& Sphere, VSREAL fOBBParameter[3], Matrix::Math::Vector3& SpherePoint) const
+{
+	return Sphere.Distance(*this, SpherePoint, fOBBParameter);
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::Distance(const Plane3& Plane, Matrix::Math::Vector3& OBBPoint, Matrix::Math::Vector3& PlanePoint) const
+{
+	return Plane.Distance(*this, PlanePoint, OBBPoint);
+}
+/*----------------------------------------------------------------*/
+VSREAL OBB3::SquaredDistance(const Polygon3& Polygon,
+	VSREAL OBBParameter[3], int& IndexTriangle,
+	VSREAL TriangleParameter[3]) const
+{
+	return Polygon.SquaredDistance(*this, IndexTriangle, TriangleParameter, OBBParameter);
+}
+
+
+
+int OBB3::RelationWith(const Matrix::Math::Vector3& Point) const
+{
+	return Point.RelationWith(*this);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Line3& Line, unsigned int& Quantity, VSREAL& tNear, VSREAL& tFar) const
+{
+
+	return Line.RelationWith(*this, Quantity, tNear, tFar);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Ray3& Ray, unsigned int& Quantity, VSREAL& tNear, VSREAL& tFar) const
+{
+
+	return Ray.RelationWith(*this, Quantity, tNear, tFar);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Segment3& Segment, unsigned int& Quantity, VSREAL& tNear, VSREAL& tFar) const
+{
+	return Segment.RelationWith(*this, Quantity, tNear, tFar);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Plane3& Plane) const
+{
+	Matrix::Math::Vector3 N = Plane.GetN();
+	VSREAL fD = Plane.GetfD();
+	VSREAL fRadius = ABS(m_fA[0] * (N.Dot(m_A[0]))) + ABS(m_fA[1] * (N.Dot(m_A[1]))) + ABS(m_fA[2] * (N.Dot(m_A[2])));
+
+	VSREAL fTest = N.Dot(m_Center) + fD;
+
+	if (fTest > fRadius)
+		return IT_Front;
+	else if (fTest < -fRadius)
+		return IT_Back;
+	else
+		return IT_Intersect;
+}
+
+int OBB3::RelationWith(const Triangle3& Triangle) const
+{
+	return Triangle.RelationWith(*this);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Rectangle3& Rectangle) const
+{
+	return Rectangle.RelationWith(*this);
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const OBB3& OBB) const
+{
+	VSREAL T[3];
+
+	Matrix::Math::Vector3 vcD = OBB.m_Center - m_Center;
+
+	VSREAL matM[3][3];
+	VSREAL ra,
+		rb,
+		t;
+
+	matM[0][0] = m_A[0].Dot(OBB.m_A[0]);
+	matM[0][1] = m_A[0].Dot(OBB.m_A[1]);
+	matM[0][2] = m_A[0].Dot(OBB.m_A[2]);
+	ra = m_fA[0];
+	rb = OBB.m_fA[0] * ABS(matM[0][0]) +
+		OBB.m_fA[1] * ABS(matM[0][1]) +
+		OBB.m_fA[2] * ABS(matM[0][2]);
+
+	T[0] = vcD.Dot(m_A[0]);
+	t = ABS(T[0]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	matM[1][0] = m_A[1].Dot(OBB.m_A[0]);
+	matM[1][1] = m_A[1].Dot(OBB.m_A[1]);
+	matM[1][2] = m_A[1].Dot(OBB.m_A[2]);
+	ra = m_fA[1];
+	rb = OBB.m_fA[0] * ABS(matM[1][0]) +
+		OBB.m_fA[1] * ABS(matM[1][1]) +
+		OBB.m_fA[2] * ABS(matM[1][2]);
+	T[1] = vcD.Dot(m_A[1]);
+	t = ABS(T[1]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	matM[2][0] = m_A[2].Dot(OBB.m_A[0]);
+	matM[2][1] = m_A[2].Dot(OBB.m_A[1]);
+	matM[2][2] = m_A[2].Dot(OBB.m_A[2]);
+	ra = m_fA[2];
+	rb = OBB.m_fA[0] * ABS(matM[2][0]) +
+		OBB.m_fA[1] * ABS(matM[2][1]) +
+		OBB.m_fA[2] * ABS(matM[2][2]);
+	T[2] = vcD.Dot(m_A[2]);
+	t = ABS(T[2]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[0][0]) +
+		m_fA[1] * ABS(matM[1][0]) +
+		m_fA[2] * ABS(matM[2][0]);
+	rb = OBB.m_fA[0];
+	t = ABS(T[0] * matM[0][0] + T[1] * matM[1][0] + T[2] * matM[2][0]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[0][1]) +
+		m_fA[1] * ABS(matM[1][1]) +
+		m_fA[2] * ABS(matM[2][1]);
+	rb = OBB.m_fA[1];
+	t = ABS(T[0] * matM[0][1] + T[1] * matM[1][1] + T[2] * matM[2][1]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[0][2]) +
+		m_fA[1] * ABS(matM[1][2]) +
+		m_fA[2] * ABS(matM[2][2]);
+	rb = OBB.m_fA[2];
+	t = ABS(T[0] * matM[0][2] + T[1] * matM[1][2] + T[2] * matM[2][2]);
+	if (t > (ra + rb))
+		return IT_NoIntersect;
+
+	ra = m_fA[1] * ABS(matM[2][0]) + m_fA[2] * ABS(matM[1][0]);
+	rb = OBB.m_fA[1] * ABS(matM[0][2]) + OBB.m_fA[2] * ABS(matM[0][1]);
+	t = ABS(T[2] * matM[1][0] - T[1] * matM[2][0]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[1] * ABS(matM[2][1]) + m_fA[2] * ABS(matM[1][1]);
+	rb = OBB.m_fA[0] * ABS(matM[0][2]) + OBB.m_fA[2] * ABS(matM[0][0]);
+	t = ABS(T[2] * matM[1][1] - T[1] * matM[2][1]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[1] * ABS(matM[2][2]) + m_fA[2] * ABS(matM[1][2]);
+	rb = OBB.m_fA[0] * ABS(matM[0][1]) + OBB.m_fA[1] * ABS(matM[0][0]);
+	t = ABS(T[2] * matM[1][2] - T[1] * matM[2][2]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[2][0]) + m_fA[2] * ABS(matM[0][0]);
+	rb = OBB.m_fA[1] * ABS(matM[1][2]) + OBB.m_fA[2] * ABS(matM[1][1]);
+	t = ABS(T[0] * matM[2][0] - T[2] * matM[0][0]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[2][1]) + m_fA[2] * ABS(matM[0][1]);
+	rb = OBB.m_fA[0] * ABS(matM[1][2]) + OBB.m_fA[2] * ABS(matM[1][0]);
+	t = ABS(T[0] * matM[2][1] - T[2] * matM[0][1]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[2][2]) + m_fA[2] * ABS(matM[0][2]);
+	rb = OBB.m_fA[0] * ABS(matM[1][1]) + OBB.m_fA[1] * ABS(matM[1][0]);
+	t = ABS(T[0] * matM[2][2] - T[2] * matM[0][2]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[1][0]) + m_fA[1] * ABS(matM[0][0]);
+	rb = OBB.m_fA[1] * ABS(matM[2][2]) + OBB.m_fA[2] * ABS(matM[2][1]);
+	t = ABS(T[1] * matM[0][0] - T[0] * matM[1][0]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[1][1]) + m_fA[1] * ABS(matM[0][1]);
+	rb = OBB.m_fA[0] * ABS(matM[2][2]) + OBB.m_fA[2] * ABS(matM[2][0]);
+	t = ABS(T[1] * matM[0][1] - T[0] * matM[1][1]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	ra = m_fA[0] * ABS(matM[1][2]) + m_fA[1] * ABS(matM[0][2]);
+	rb = OBB.m_fA[0] * ABS(matM[2][1]) + OBB.m_fA[1] * ABS(matM[2][0]);
+	t = ABS(T[1] * matM[0][2] - T[0] * matM[1][2]);
+	if (t > ra + rb)
+		return IT_NoIntersect;
+
+	return IT_Intersect;
+}
+/*----------------------------------------------------------------*/
+int OBB3::RelationWith(const Sphere3& Sphere) const
+{
+	VSREAL fOBBParameter[3];
+	VSREAL Dist = Sphere.m_Center.SquaredDistance(*this, fOBBParameter);
+	Dist = SQRT(Dist);
+	if (Dist - Sphere.m_fRadius < 0.0f)
+		return IT_Intersect;
+
+	return IT_NoIntersect;
+}
