+++ Engine/Source/Runtime/Primitive/Plane3.h
@@ -39,73 +39,73 @@ namespace Matrix
 			Matrix::Math::Vector3 ReflectDir(const Matrix::Math::Vector3& Dir) const;
 			void GetReverse(Plane3& OutPlane) const;
 
-			///*************************************距离************************************************/
-			////点到平面距离
-			// VSREAL Distance(const Matrix::Math::Vector3& Point, Matrix::Math::Vector3& PlanePoint)const;
-			////平面和球的距离
-			// VSREAL Distance(const Sphere3& Sphere, Matrix::Math::Vector3& SpherePoint)const;
-			////直线和平面距离
-			// VSREAL Distance(const Line3& Line, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& LinePoint)const;
-			////射线和平面距离
-			// VSREAL Distance(const Ray3& Ray, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& RayPoint)const;
-			////线段和平面距离
-			// VSREAL Distance(const Segment3& Segment, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& SegmentPoint)const;
+			/*************************************距离************************************************/
+			//点到平面距离
+			VSREAL Distance(const Matrix::Math::Vector3& Point, Matrix::Math::Vector3& PlanePoint)const;
+			//平面和球的距离
+			VSREAL Distance(const Sphere3& Sphere, Matrix::Math::Vector3& SpherePoint)const;
+			//直线和平面距离
+			VSREAL Distance(const Line3& Line, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& LinePoint)const;
+			//射线和平面距离
+			VSREAL Distance(const Ray3& Ray, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& RayPoint)const;
+			//线段和平面距离
+			VSREAL Distance(const Segment3& Segment, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& SegmentPoint)const;
 
-			////下面距离函数再没有考虑相交的情况下计算的，因为相交的情况下举例为0，先用RelationWith判断位置关系，再用下面来计算距离
-			////平面和平面距离
-			// VSREAL Distance(const Plane3& Plane, Matrix::Math::Vector3& Plane1Point, Matrix::Math::Vector3& Plane2Point)const;
-			////平面和三角形距离
-			// VSREAL Distance(const Triangle3& Triangle, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& TrianglePoint)const;
-			////矩形和平面距离
-			// VSREAL Distance(const Rectangle3& Rectangle, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& RectanglePoint)const;
-			////OBB和平面距离
-			// VSREAL Distance(const OBB3& OBB, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& OBBPoint)const;
-			////AABB和平面距离
-			// VSREAL Distance(const AABB3& AABB, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& AABBPoint)const;
-			////平面和多边形距离
-			// VSREAL Distance(const Polygon3& Polygon, Matrix::Math::Vector3& PlanePoint, int& IndexTriangle,
-			//	Matrix::Math::Vector3& TrianglePoint)const;
-			///********************************RelationWith******************************************/
-			////点和平面的位置关系(VSFRONT VSBACK VSPLANAR)
-			// int RelationWith(const Matrix::Math::Vector3& Point)const;
-			////测试直线与平面位置关系
-			////VSNOINTERSECT VSNTERSECT VSON VSBACK VSFRONT
-			// int RelationWith(const Line3& Line, bool bCull, VSREAL& fLineParameter)const;
-			////测试射线与平面位置关系
-			////VSNOINTERSECT VSNTERSECT VSON VSBACK VSFRONT
-			// int RelationWith(const Ray3& Ray, bool bCull, VSREAL& fRayParameter)const;
-			////测试线段与平面位置关系
-			////VSNOINTERSECT VSNTERSECT VSON VSBACK VSFRONT
-			// int RelationWith(const Segment3& Segment, bool bCull, VSREAL& fSegmentParameter)const;
-			////测试平面和OBB位置关系
-			////VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const OBB3& OBB)const;
-			////测试平面和AABB位置关系
-			////VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const AABB3& AABB)const;
-			////平面与球的位置关系
-			////VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const Sphere3& Sphere)const;
-			////平面和三角形位置关系
-			////VSON VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const Triangle3& Triangle)const;
-			// int RelationWith(const Triangle3& Triangle, Segment3& Segment)const;
-			////测试做为参数平面和平面位置关系
-			////VSNOINTERSECT VSINTERSECT
-			// int RelationWith(const Plane3& Plane)const;
-			// int RelationWith(const Plane3& Plane, Line3& Line)const;
-			////平面和矩形位置关系
-			////VSON VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const Rectangle3& Rectangle)const;
-			// int RelationWith(const Rectangle3& Rectangle, Segment3& Segment)const;
+			//下面距离函数再没有考虑相交的情况下计算的，因为相交的情况下举例为0，先用RelationWith判断位置关系，再用下面来计算距离
+			//平面和平面距离
+			VSREAL Distance(const Plane3& Plane, Matrix::Math::Vector3& Plane1Point, Matrix::Math::Vector3& Plane2Point)const;
+			//平面和三角形距离
+			VSREAL Distance(const Triangle3& Triangle, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& TrianglePoint)const;
+			//矩形和平面距离
+			VSREAL Distance(const Rectangle3& Rectangle, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& RectanglePoint)const;
+			//OBB和平面距离
+			VSREAL Distance(const OBB3& OBB, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& OBBPoint)const;
+			//AABB和平面距离
+			VSREAL Distance(const AABB3& AABB, Matrix::Math::Vector3& PlanePoint, Matrix::Math::Vector3& AABBPoint)const;
+			//平面和多边形距离
+			VSREAL Distance(const Polygon3& Polygon, Matrix::Math::Vector3& PlanePoint, int& IndexTriangle,
+				Matrix::Math::Vector3& TrianglePoint)const;
+			/********************************RelationWith******************************************/
+			//点和平面的位置关系(IT_Front IT_Back VSPLANAR)
+			int RelationWith(const Matrix::Math::Vector3& Point)const;
+			//测试直线与平面位置关系
+			//IT_NoIntersect VSNTERSECT IT_On IT_Back IT_Front
+			int RelationWith(const Line3& Line, bool bCull, VSREAL& fLineParameter)const;
+			//测试射线与平面位置关系
+			//IT_NoIntersect VSNTERSECT IT_On IT_Back IT_Front
+			int RelationWith(const Ray3& Ray, bool bCull, VSREAL& fRayParameter)const;
+			//测试线段与平面位置关系
+			//IT_NoIntersect VSNTERSECT IT_On IT_Back IT_Front
+			int RelationWith(const Segment3& Segment, bool bCull, VSREAL& fSegmentParameter)const;
+			//测试平面和OBB位置关系
+			//IT_Front IT_Back IT_Intersect
+			int RelationWith(const OBB3& OBB)const;
+			//测试平面和AABB位置关系
+			//IT_Front IT_Back IT_Intersect
+			int RelationWith(const AABB3& AABB)const;
+			//平面与球的位置关系
+			//IT_Front IT_Back IT_Intersect
+			int RelationWith(const Sphere3& Sphere)const;
+			//平面和三角形位置关系
+			//IT_On IT_Front IT_Back IT_Intersect
+			int RelationWith(const Triangle3& Triangle)const;
+			int RelationWith(const Triangle3& Triangle, Segment3& Segment)const;
+			//测试做为参数平面和平面位置关系
+			//IT_NoIntersect IT_Intersect
+			int RelationWith(const Plane3& Plane)const;
+			int RelationWith(const Plane3& Plane, Line3& Line)const;
+			//平面和矩形位置关系
+			//IT_On IT_Front IT_Back IT_Intersect
+			int RelationWith(const Rectangle3& Rectangle)const;
+			int RelationWith(const Rectangle3& Rectangle, Segment3& Segment)const;
 
-			////平面和多边形位置关系
-			////VSON VSFRONT VSBACK VSINTERSECT
-			// int RelationWith(const Polygon3& Polygon)const;
-			// int RelationWith(const Polygon3& Polygon, Segment3& Segment)const;
+			//平面和多边形位置关系
+			//IT_On IT_Front IT_Back IT_Intersect
+			int RelationWith(const Polygon3& Polygon)const;
+			int RelationWith(const Polygon3& Polygon, Segment3& Segment)const;
 
-			////平面和圆柱位置关系
-			// int RelationWith(const VSCylinder3& Cylinder3)const;
+			//平面和圆柱位置关系
+			int RelationWith(const VSCylinder3& Cylinder3)const;
 
 		}; // class
 		   /*----------------------------------------------------------------*/
